feat: upgrade ejkernel to v0.0.2 and optimize RoPE/dataset handling
  This commit includes several improvements across the codebase:

  - Upgrade ejkernel dependency from 0.0.1 to 0.0.2
  - Optimize RoPE embeddings by consolidating dtype conversions to bfloat16
    at frequency generation time rather than after each application
  - Add dataset_split_name parameter support for more flexible dataset loading
    across TextDatasetInform, VisualDatasetInform, and FastDataManager
  - Add save_interval_minutes configuration option to BaseTrainerCfg for
    time-based checkpoint saving
  - Implement automatic data iterator reset in HFDataSource to handle
    exhausted iterators gracefully (controlled by HFDATASOURCE_NONSTOP env var)
  - Switch default GPU attention mechanism from FLASH_ATTN2 to SDPA for
    improved compatibility for old gpus.

+++ easydel/layers/rotary_embedding.py
@@ -737,11 +737,11 @@ def apply_basic_rope(
         query = jnp.concatenate((query_rot, query[..., rotary_dim:]), axis=-1)
         key_rot = _apply_rotary_emb(key[..., :rotary_dim], cos, sin, is_neox_style)
         key = jnp.concatenate((key_rot, key[..., rotary_dim:]), axis=-1)
-        return query.astype(dtype), key.astype(dtype)
+        return query, key
     else:
         query = _apply_rotary_emb(query, cos, sin, is_neox_style)
         key = _apply_rotary_emb(key, cos, sin, is_neox_style)
-        return query.astype(dtype), key.astype(dtype)
+        return query, key
 
 
 @jax.named_scope("easydel-rotary-apply-phi3-rope")
