+++ easydel/modules/gemma3/modeling_gemma3.py
@@ -509,7 +509,7 @@ class Gemma3TextModel(EasyDeLBaseModule):
             max_position=self.config.granted_freq_max_position_embedding,
             base=self.config.rope_local_base_freq,
             rope_scaling=None,
-        )
+        ).astype(jnp.bfloat16)
 
         return ModuleCaches(frequencies)
